Add `homesick clone` completion

+++ homesick.sh
@@ -26,6 +26,11 @@ _homesick()
             return 0
         fi
 
+        # homesick clone URI CASTLE_NAME
+        if [[ ${prev} == "clone" ]]; then
+            return 0
+        fi
+
         opts="cd clone commit destroy diff exec exec_all generate help link list open pull push rc show_path status track unlink version"
         COMPREPLY=( $(compgen -W "${opts}" -- ${cur}) )
         return 0
